Files organization
Changed the files to remove the output files from the git folder: since
they change often, is cumbersome to keep having commits with changes in
those file (that are irrelevant). Now output goes to a folder in the
Thesis folder (backed up in dropbox). Changes to Netlogo and Python
code in order to update the folder paths accordingly