Joint machines for signal completed (separate file)
Joint machines and minimisation completed for the with signal file.
Added a procedure in ‘minimisation’ to create the joint machines with
signal, and then normal canonical form and minimisation is done (joint
machines with signal are normal machines). The whole file was
completed, skipping the parts that required the Unused states measures.