Code organisation and parents of joint machine tracking
Avoided previous use of “iterrows” in joint machine, used states and
regime procedures. This was done so parents of joint machine can be
track. Not only should make code faster (iterrows is supposedly very
slow) but makes it much easier to understand. Keeping track of parent
machines and separating used states procedure can be useful when
implementing the used states measure with signal.