Several changes, not major.
Included for both epochs and freqj, both signal and no signal, a
clearer printed view of the machines, including the cycle (for pdf,
states with p>0.1). Included time printing before saving output to
calculate times or runs. Put at the beginning globals for controlling
the pdf creation, as well as for epoch and regime classification