Individual machines analysis
A new file is included with a script for visualising the top individual
machines per epoch for each different CEM value. It was done by
printing a file with all of the minimized machines for the superlong
run with signal. Small changes to the main and graphs have been done
(since writing of the chapter required small tests and few things, but
there’s nothing significant)